Refined examples.
Removed unneeded life_keeper parameter.
In the publish handler, the buffer has itself lifetime.

Fixed retain handling on MQTT v3.1.1.
http://docs.oasis-open.org/mqtt/mqtt/v3.1.1/os/mqtt-v3.1.1-os.html#_Toc398718038
3.3.1.3 RETAIN
It MUST set the RETAIN flag to 0 when a PUBLISH Packet is sent to a
Client because it matches an established subscription regardless of
how the flag was set in the message it received [MQTT-3.3.1-9].

Added retain handling on MQTT v5.
https://docs.oasis-open.org/mqtt/mqtt/v5.0/os/mqtt-v5.0-os.html#_Toc3901104
If the value of Retain As Published subscription option is set to 0,
the Server MUST set the RETAIN flag to 0 when forwarding an
Application Message regardless of how the RETAIN flag was set in the
received PUBLISH packet [MQTT-3.3.1-12].

If the value of Retain As Published subscription option is set to 1,
the Server MUST set the RETAIN flag equal to the RETAIN flag in the
received PUBLISH packet [MQTT-3.3.1-13].