Fixed TopicAlias functionality.
The the TopicAlias map should be cleared on disconnection.
If the inflight message has the TopicAlias property, remove it.
If the inflight message has the empty TopicName and the TopicAlias,
the original TopicName is recovered from the TopicAlias at the first
sent, and remove TopicAlias property.

It works well with mosquitto MQTT broker.
The original mqtt_cpp implementation cause protocol error because of
using cleared topic alias on resend.