Use SYSTEM in target_include_directories()
This allows one to use/compile this library in a project where stricter compiler settings
(like `-pedantic` and `-Werror`) are in use. *gcc* and *clang* ignores such issues in all
system includes and as a consequence of adding `SYSTEM` to
`target_include_directories()`, the headers of this library are treated as system includes
as well.