Fixed broker side qos2 exactly once logic.
endpoint has qos2 duplicated transfer avoiding mechanism.
It is implemented by `qos2_publish_handled_`.

endpoint works for both client and broker.

However broker (server) is different from the client.
Client can reconnect usign the same endpoint instance as the previous
connection. But the broker can't.
It is because accepting mechanism. On accept, the server needs to create
the new endpoint.

So the broker had an original qos2 duplicated transfer avoiding
mechanism.
It is implemented by `qos2_publish_processed_`.

But the implementation is a little redundant.
So I introduced get and restore functionality for
`qos2_publish_handled_`.
It was #873 but it had bug. `qos2_publish_processed_` was mixed.
This PR completely removed `qos2_publish_processed_` related code from
the broker and added correct implementation for `qos2_publish_handled_`.