Fixed qos2_publish_handled_ invalidly remains issue.
qos2_publish_handled_ is used for avoid qos2 publish handling twice or
more. On publish receive, if it doesn't contain the packet_id, then handle the
publish message and add the packet_id to qos2_publish_handled_.
On pubrel receive, erase the packet_id from qos2_publish_handled_
because the publish packet that has the packet_id is no logner resend.
It is the basic logic.
However, when user sends pubrec with error reason_code, then the communication
partner doesn't send back pubrel due to the error.
In this case, there is no chance to erase qos2_publish_handled_ entry.

It should be erased when user sends pubrec with error reason_code.
This fix adds it.

qos2_publish_handled_ used to access from only on strand. But this
addition introduce out of strand access. It is user's timing. So I
introduced mutex for qos2_publish_handled_.